Hide or remove several members in package `scala.concurrent` which should be private.
- hide object `concurrent.impl.Promise`
- move `isFutureThrowable` to `concurrent.Future` trait
- remove `concurrent.ConcurrentPackageObject`
- remove `concurrent.handledFutureException` member
- hide `concurrent.currentExecutionContext`
